Codecs are dead. Long live Conversions.
The things previously called Codecs are now dead.
After playing with the interface a lot and writing several
transformer modules, it became clear that the Codec/Conversion
abstraction is more problematic than helpful. Since all Codecs
can be seen as just Conversions, this works out fine.

This commit makes significant changes to how the modules work.
Basically now there are only `Types` and `Conversions`.

  `Types` identify a kind of data.
  `Conversions` identify a function to convert between `Types`.

As of now, conversions and their inversions are written in
separate modules. This is a bit annoying, but is due to how
modules are meant to be loaded. Already exploring symmetric
loading (if `transformer-a-to-b` is not there, try
`transformer-b-to-a` and check for an `inverse` property). This
is also annoying for its own reasons (ambiguity when calling it,
ordering conventions, etc).

The good news:

  > echo '127.0.0.1' | transform ip-address buffer hex
  7f000001
  > echo '7f000001' | transform hex buffer ip-address
  127.0.0.1

Boom.